refactor: remove Stripe references from app backend

+++ convex/app.ts
@@ -1,7 +1,4 @@
-import { internal } from "@cvx/_generated/api";
 import { mutation, query } from "@cvx/_generated/server";
-import { currencyValidator, PLANS } from "@cvx/schema";
-import { asyncMap } from "convex-helpers";
 import { v } from "convex/values";
 import { User } from "~/types";
 import { getUserId, ensureUser } from "@cvx/auth";
@@ -20,25 +17,16 @@ export const getCurrentUser = query({
 
     const subscription = await ctx.db
       .query("subscriptions")
-      .withIndex("userId", (q) => q.eq("userId", user._id))
-      .unique();
+      .withIndex("userId", (q) => q.eq("userId", identity.subject))
+      .first();
 
-    const plan = subscription?.planId
-      ? await ctx.db.get(subscription.planId)
-      : undefined;
     const avatarUrl = user.imageId
       ? await ctx.storage.getUrl(user.imageId)
       : user.image;
     return {
       ...user,
       avatarUrl: avatarUrl || undefined,
-      subscription:
-        subscription && plan
-          ? {
-              ...subscription,
-              planKey: plan.key,
-            }
-          : undefined,
+      subscription: subscription ?? undefined,
     };
   },
 });
@@ -68,23 +56,11 @@ export const updateUsername = mutation({
 export const completeOnboarding = mutation({
   args: {
     username: v.string(),
-    currency: currencyValidator,
   },
   handler: async (ctx, args) => {
     const user = await ensureUser(ctx);
     if (!user) return;
-
     await ctx.db.patch(user._id, { username: args.username });
-    if (user.customerId) return;
-
-    await ctx.scheduler.runAfter(
-      0,
-      internal.stripe.PREAUTH_createStripeCustomer,
-      {
-        currency: args.currency,
-        userId: user._id,
-      },
-    );
   },
 });
 
@@ -117,27 +93,6 @@ export const removeUserImage = mutation({
   },
 });
 
-export const getActivePlans = query({
-  args: {},
-  handler: async (ctx) => {
-    const identity = await ctx.auth.getUserIdentity();
-    if (!identity) return;
-
-    const [free, pro] = await asyncMap(
-      [PLANS.FREE, PLANS.PRO] as const,
-      (key) =>
-        ctx.db
-          .query("plans")
-          .withIndex("key", (q) => q.eq("key", key))
-          .unique(),
-    );
-    if (!free || !pro) {
-      throw new Error("Plan not found");
-    }
-    return { free, pro };
-  },
-});
-
 export const deleteCurrentUserAccount = mutation({
   args: {},
   handler: async (ctx) => {
@@ -147,18 +102,15 @@ export const deleteCurrentUserAccount = mutation({
     const user = await ctx.db.get(userId);
     if (!user) throw new Error("User not found");
 
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) throw new Error("Unauthorized");
+
     const subscription = await ctx.db
       .query("subscriptions")
-      .withIndex("userId", (q) => q.eq("userId", userId))
-      .unique();
-    if (!subscription) {
-      console.error("No subscription found");
-    } else {
+      .withIndex("userId", (q) => q.eq("userId", identity.subject))
+      .first();
+    if (subscription) {
       await ctx.db.delete(subscription._id);
-      await ctx.scheduler.runAfter(
-        0,
-        internal.stripe.cancelCurrentUserSubscriptions,
-      );
     }
     await ctx.db.delete(userId);
   },
